Reverting previous changes
Turns out it has something to do with the SkillsArticle or even the ArticleGrid that isn't displaying the svg. I'm reverting all the way back to passing in an array of logo strings, but I will look into this.
As of right now, skills aren't being properly displayed on mobile.